新增购课的关联处理: 1、更新member_course表 2、更新教练日计表（t_coach_lesson_reg） 3、更新场地日计表（t_club_lesson_reg)

+++ src/main/java/com/atguigu/controller/CoachScheduleController.java
@@ -29,7 +29,7 @@ public class CoachScheduleController {
                                 @RequestParam("real_club") String real_club,
                                 @RequestParam("real_coach") String real_coach,
                                 @RequestParam("sale_id") String sale_id,
-                                @RequestParam("seq_no") Byte seq_no,
+                                @RequestParam("seq_no") Integer seq_no,
                                 @RequestParam("start_time_2") String start_time,
                                 @RequestParam("end_time_2") String end_time,
                                 @RequestParam("bz1") String bz1,
