+++ src/main/java/com/atguigu/service/CoachScheduleService.java
@@ -4,6 +4,6 @@ import java.util.Date;
 
 public interface CoachScheduleService {
     public boolean updatelesson(String mem_id, String real_club, String real_coach,
-                                String sale_id, Byte seq_no, Date start_time_2,
+                                String sale_id, Integer seq_no, Date start_time_2,
                                 Date end_time_2, String bz1);
 }
